# Response.content minor perf improvement