Python 2
```console
$ python -m timeit -s "blist = [b'foo', b'bar'];" "bytes().join(blist)"
1000000 loops, best of 3: 0.243 usec per loop
$ python -m timeit -s "blist = [b'foo', b'bar'];" "b''.join(blist)"
10000000 loops, best of 3: 0.125 usec per loop
```

Python 3
```console
$ python -m timeit -s "blist = [b'foo', b'bar'];" "bytes().join(blist)"
1000000 loops, best of 3: 0.285 usec per loop
$ python -m timeit -s "blist = [b'foo', b'bar'];" "b''.join(blist)"
10000000 loops, best of 3: 0.161 usec per loop
```