@sigmavirus24 2018-05-27
Thanks for catching this @ofek!

If others are curious as to why this is happening it's because the syntactic sugar for `b''` built into the language is faster than creating an object via `bytes()`. The same actually holds true for `[]` instead of `list()`, `{}` instead of `dict()`, and the list goes on.